feat: New task documentation method. Closes #14
Annotating a task with a documentation string has been improved. Now,
this is done with a comment _before_ the function name, instead of on
the same line as the function. The old way has been preserved for
backwards compatability. However, a warning message is printed with
instructions to migrate to the new method.