perf: Minimize extraneous sources
Previously, under watch mode, the Bakefile would be sourced _twice_, and parsed
through line-by-line _three times_ to see if a particular task exists (and to
read potential watch exec arguments). Now, this is only done once and
twice, respectively, after watchexec is started with the watchfile
arguments. In circumstances such as passing incorrect task names,
execution might take a bit longer because it execs into 'watchexec'
before spitting out an error. But, that's a non-happy path and the code
still looks nicer.