Redo breaking change of removing `bake.assert_nonempty`
Originally, I was going to approach the new release by undoing the
breaking changes on `HEAD` and releasing a minor version. And then
redoing the breaking changes to then release a `v2`.

That's why I made this commit.

But I changed my mind, and now I'm simply going to work towards a `v2` (with
some prereleases).