Improve detection of `Bakefile.sh` to search in same dir as `bake`
If the `Bakefile.sh` cannot be found by walking up the parent
directories, it will search for it in the same directory as `bake`.

This can be useful when working on a foreign project (where you may
want to run various tasks, but not commit anything to version control).
Say, if `.hidden/` is in the global gitignore, all the bake stuff can be
placed in there, and bake will now find the Bakefile.